Minor changes, major mistakes will be corrected lately
