+++ src/main/java/ru/sbt/mipt/oop/AllDoorsCloseScenario.java
@@ -9,7 +9,7 @@ class AllDoorsCloseScenario implements Scenarios {
                     for (Light light : homeRoom.getLights()) {
                         light.setOn(false);
                         SensorCommand command = new SensorCommand(CommandType.LIGHT_OFF, light.getId());
-                        EventManager.sendCommand(command);
+                        CommandManager.sendCommand(command);
                     }
                 }
             }
