+++ src/main/java/ru/sbt/mipt/oop/CommandManager.java
@@ -1,0 +1,7 @@
+package ru.sbt.mipt.oop;
+
+public class CommandManager {
+    static void sendCommand(SensorCommand command) {
+        System.out.println("Pretent we're sending command " + command);
+    }
+}
