--- src/main/java/ru/sbt/mipt/oop/EventHandler.java
@@ -1,5 +1,0 @@
-package ru.sbt.mipt.oop;
-
-interface EventHandler {
-    void run(SensorEventType type, String objectId, SmartHome smartHome);
-}
