--- src/main/java/ru/sbt/mipt/oop/HomeCondition.java
@@ -1,7 +1,0 @@
-package ru.sbt.mipt.oop;
-
-import java.io.IOException;
-
-public interface HomeCondition {
-     SmartHome readCondition() throws IOException;
-}
