--- src/main/java/ru/sbt/mipt/oop/HomeConditionFromFile.java
@@ -1,17 +1,0 @@
-package ru.sbt.mipt.oop;
-
-import com.google.gson.Gson;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-class HomeConditionFromFile implements HomeCondition {
-    @Override
-    public SmartHome readCondition() throws IOException {
-        Gson gson = new Gson();
-        String json = new String(Files.readAllBytes(Paths.get("smart-home-1.js")));
-        SmartHome smartHome = gson.fromJson(json, SmartHome.class);
-        return smartHome;
-    }
-}
