+++ src/main/java/ru/sbt/mipt/oop/ReadHomeState.java
@@ -1,0 +1,7 @@
+package ru.sbt.mipt.oop;
+
+import java.io.IOException;
+
+public interface ReadHomeState {
+     SmartHome readCondition();
+}
